\name{ebar.spp}
\docType{data}
\alias{ebar.spp}
\title{Early Barstovian Locality Data}
\description{
  This data set provides the taxonomic occurrence data for all the Early Barstovian localities from the Miocene Mammal Mapping Project (MIOMAP).
}
\usage{ebar.spp}
\format{A data frame containing 1645 taxa occurrences (rows) with data in 16 different fields (columns).}
\source{MIOMAP project}
\references{
Carrasco, M., Kraatz, B. Davis, E. and Barnosky., A. (2005) \emph{Miocene Mammal Mapping Project (MIOMAP)}. University of California Museum of Paleontology.

Carrasco, M., Barnosky, A., Kraatz, B. and Davis, E. (2007) \emph{The Miocene Mammal Mapping Project (MIOMAP): An Online Database of Arikareean Through Hemphillian Fossil Mammals}. Bulletin of Carnegie Museum of Natural History, 39: 183--188
}
\keyword{datasets}

