\name{fossil-package}
\alias{fossil-package}
\alias{fossil}
\docType{package}
\title{
fossil: Palaeoecological and Palaeogeographical Analysis Tools
}
\description{
A set of analytical tools useful in analysing ecological and geographical data sets, both ancient and modern. The package includes functions for estimating species richness (Chao 1 and 2, ACE, ICE, Jacknife), shared species/beta diversity, species area curves and geographic distances.
}
\details{The fossil package is designed to be used by palaeoecologists and palaeobiogeographers, providing to them a set of useful tools including species similarity indices, species estimators, minimum spanning tree/forest functions, and an assortment of other useful tools.}
\author{
Matthew Vavrek <matthew@matthewvavrek.com>
}
\keyword{ package }

