\name{sppexample}
\alias{sppexample}
\title{A Sample Species Abundance Dataset}
\description{
A simple hypothetical data set used in many of the examples.
}
\value{
A matrix with 12 samples with a total of 12 different unique species. Abundances are between 1 and 10.  
}
\author{Matthew Vavrek}
\examples{
data(sppexample)
}
\keyword{data}
