\name{fossil-package}
\alias{fossil-package}
\alias{fossil}
\docType{package}
\title{
Palaeoecological and Palaeogeographical Analysis Tools
fossil
}
\description{
A set of analytical tools useful in analysing 
	ecological and geographical data sets, both ancient and modern. 
  The package includes functions for estimating species richness 
  (Chao 1	and 2, ACE, ICE, Jacknife), shared species/beta diversity,
	species area curves and geographic distances.
}
\details{
\tabular{ll}{
Package: \tab fossil\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2007-11-21\cr
License: \tab GPL version 2 or newer\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Matthew Vavrek <matthew.vavrek@gmail.com>

Maintainer: Matthew Vavrek <matthew.vavrek@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-shapefiles]{<pkg>}} ~~
}
\examples{
##add examples
}
